feat: use runtime.AddCleanup for BufParams
